requirements.txt
beautifulsoup4==4.12.2
pandas==2.1.4
lxml==4.9.3
requests
openai>=1.0.0
pymongo>=4.0.0
motor>=3.0.0
python-dotenv>=1.0.0
fastapi>=0.116.1
uvicorn>=0.35.0